requirements.txt
streamlit
openai>=1.12.0
python-dotenv
gspread
oauth2client